refactor(domain): decompose card_commands.rs into commands/card submodules (KAN-768)
