+++ crates/kanban-domain/src/commands/card/lifecycle.rs
@@ -1,0 +1,346 @@
+use super::super::{Command, CommandContext};
+use super::CardCommand;
+use crate::data_store::DataStore;
+use crate::{CardUpdate, CreateCardOptions, DomainError, KanbanError, KanbanResult};
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+/// Update card properties (title, description, priority, status, etc.)
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct UpdateCard {
+    pub card_id: Uuid,
+    pub updates: CardUpdate,
+}
+
+impl UpdateCard {
+    pub fn execute(&self, context: &CommandContext) -> KanbanResult<()> {
+        let mut card = context.get_card(self.card_id)?;
+        card.update(self.updates.clone(), Utc::now());
+        context.store.upsert_card(card)?;
+        Ok(())
+    }
+
+    pub fn description(&self) -> String {
+        "Update card".to_string()
+    }
+
+    /// Inverse: read the card's current state and synthesise an
+    /// `UpdateCard` whose `updates` field-by-field restore each touched
+    /// field to its prior value. Fields not touched by the forward
+    /// command stay `None` / `NoChange` so the inverse is minimal.
+    pub fn capture_inverse(&self, store: &dyn DataStore) -> KanbanResult<Vec<Command>> {
+        use crate::field_update::FieldUpdate;
+        let card = match store.get_card(self.card_id)? {
+            Some(c) => c,
+            None => return Err(KanbanError::not_found("Card", self.card_id)),
+        };
+
+        let upd = &self.updates;
+        let inverse = CardUpdate {
+            title: upd.title.as_ref().map(|_| card.title.clone()),
+            description: match upd.description {
+                FieldUpdate::NoChange => FieldUpdate::NoChange,
+                _ => match card.description {
+                    Some(v) => FieldUpdate::Set(v),
+                    None => FieldUpdate::Clear,
+                },
+            },
+            priority: upd.priority.map(|_| card.priority),
+            status: upd.status.map(|_| card.status),
+            position: upd.position.map(|_| card.position),
+            column_id: upd.column_id.map(|_| card.column_id),
+            due_date: match upd.due_date {
+                FieldUpdate::NoChange => FieldUpdate::NoChange,
+                _ => match card.due_date {
+                    Some(v) => FieldUpdate::Set(v),
+                    None => FieldUpdate::Clear,
+                },
+            },
+            points: match upd.points {
+                FieldUpdate::NoChange => FieldUpdate::NoChange,
+                _ => match card.points {
+                    Some(v) => FieldUpdate::Set(v),
+                    None => FieldUpdate::Clear,
+                },
+            },
+            sprint_id: match upd.sprint_id {
+                FieldUpdate::NoChange => FieldUpdate::NoChange,
+                _ => match card.sprint_id {
+                    Some(v) => FieldUpdate::Set(v),
+                    None => FieldUpdate::Clear,
+                },
+            },
+        };
+
+        Ok(vec![Command::Card(CardCommand::Update(UpdateCard {
+            card_id: self.card_id,
+            updates: inverse,
+        }))])
+    }
+}
+
+/// Create a new card in a column
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct CreateCard {
+    pub id: Uuid,
+    pub card_number: u32,
+    pub board_id: Uuid,
+    pub column_id: Uuid,
+    pub title: String,
+    pub position: i32,
+    pub options: CreateCardOptions,
+    #[serde(default = "chrono::Utc::now")]
+    pub timestamp: DateTime<Utc>,
+}
+
+impl CreateCard {
+    pub fn execute(&self, context: &CommandContext) -> KanbanResult<()> {
+        context.check_wip_limit(self.column_id, 1, &[])?;
+        let mut board = context.get_board(self.board_id)?;
+
+        let now = self.timestamp;
+        let mut card = crate::Card {
+            id: self.id,
+            column_id: self.column_id,
+            title: self.title.clone(),
+            description: None,
+            priority: crate::CardPriority::Medium,
+            status: crate::CardStatus::Todo,
+            position: self.position,
+            due_date: None,
+            points: None,
+            card_number: self.card_number,
+            sprint_id: None,
+            created_at: now,
+            updated_at: now,
+            completed_at: None,
+            sprint_logs: Vec::new(),
+        };
+
+        if board.card_counter <= self.card_number {
+            board.card_counter = self.card_number + 1;
+        }
+
+        if self.options.description.is_some()
+            || self.options.priority.is_some()
+            || self.options.points.is_some()
+            || self.options.due_date.is_some()
+        {
+            let updates = CardUpdate {
+                description: self
+                    .options
+                    .description
+                    .clone()
+                    .map(crate::FieldUpdate::Set)
+                    .unwrap_or(crate::FieldUpdate::NoChange),
+                priority: self.options.priority,
+                points: self
+                    .options
+                    .points
+                    .map(crate::FieldUpdate::Set)
+                    .unwrap_or(crate::FieldUpdate::NoChange),
+                due_date: self
+                    .options
+                    .due_date
+                    .map(crate::FieldUpdate::Set)
+                    .unwrap_or(crate::FieldUpdate::NoChange),
+                ..Default::default()
+            };
+            card.update(updates, now);
+        }
+
+        if let Some(sprint_id) = self.options.sprint_id {
+            let sprint = context.get_sprint(sprint_id)?;
+            if sprint.board_id != self.board_id {
+                return Err(KanbanError::Domain(DomainError::SprintBoardMismatch {
+                    sprint_id,
+                    sprint_board: sprint.board_id,
+                    card_board: self.board_id,
+                }));
+            }
+            let sprint_number = sprint.sprint_number;
+            let sprint_name = sprint.get_name(&board).map(|s| s.to_string());
+            let sprint_status = format!("{:?}", sprint.status);
+            card.assign_to_sprint(sprint_id, sprint_number, sprint_name, sprint_status, now);
+        }
+
+        context.store.upsert_board(board)?;
+        context.store.upsert_card(card)?;
+        Ok(())
+    }
+
+    pub fn description(&self) -> String {
+        format!("Create card: '{}'", self.title)
+    }
+
+    /// Inverse: delete the new card. `DeleteCard` is polymorphic over
+    /// live / archived so it cleanly removes a freshly-created live
+    /// card without leaving an archive trail. The board's
+    /// `card_counter` stays bumped; redo via the original forward
+    /// reproduces the same id and number.
+    pub fn capture_inverse(&self, _store: &dyn DataStore) -> KanbanResult<Vec<Command>> {
+        Ok(vec![Command::Card(CardCommand::Delete(DeleteCard {
+            card_id: self.id,
+        }))])
+    }
+}
+
+/// Restore an archived card
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct RestoreCard {
+    pub card_id: Uuid,
+    pub column_id: Uuid,
+    pub position: i32,
+    #[serde(default = "chrono::Utc::now")]
+    pub timestamp: DateTime<Utc>,
+}
+
+impl RestoreCard {
+    /// Inverse: archive the card again. The card id is in the forward
+    /// command. ArchiveCards captures original column/position from the
+    /// live card at capture time — by the time this runs the card has
+    /// been restored to (self.column_id, self.position), so the
+    /// re-archive will use those values as the new "original" location.
+    pub fn capture_inverse(&self, _store: &dyn DataStore) -> KanbanResult<Vec<Command>> {
+        Ok(vec![Command::Card(CardCommand::Archive(ArchiveCards {
+            ids: vec![self.card_id],
+        }))])
+    }
+
+    pub fn execute(&self, context: &CommandContext) -> KanbanResult<()> {
+        context.check_wip_limit(self.column_id, 1, &[])?;
+        let archived = context
+            .store
+            .get_archived_card(self.card_id)?
+            .ok_or_else(|| KanbanError::not_found("archived card", self.card_id))?;
+        let mut card = archived.into_card();
+        card.column_id = self.column_id;
+        card.position = self.position;
+        card.updated_at = self.timestamp;
+
+        context.store.delete_archived_card(self.card_id)?;
+        context.store.upsert_card(card)?;
+
+        let card_id = self.card_id;
+        context.store.modify_graph(Box::new(move |graph| {
+            graph.unarchive_node(card_id);
+            Ok(())
+        }))?;
+        Ok(())
+    }
+
+    pub fn description(&self) -> String {
+        format!("Restore card {}", self.card_id)
+    }
+}
+
+/// Permanently delete a card. Operates on whichever list the card is
+/// in — live or archived. Strips incident graph edges.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct DeleteCard {
+    pub card_id: Uuid,
+}
+
+impl DeleteCard {
+    pub fn execute(&self, context: &CommandContext) -> KanbanResult<()> {
+        // Both store ops are idempotent on missing — calling both
+        // covers a card in either list.
+        context.store.delete_card(self.card_id)?;
+        context.store.delete_archived_card(self.card_id)?;
+        let card_id = self.card_id;
+        context.store.modify_graph(Box::new(move |graph| {
+            graph.remove_node(card_id);
+            Ok(())
+        }))?;
+        Ok(())
+    }
+
+    pub fn description(&self) -> String {
+        format!("Delete card {}", self.card_id)
+    }
+
+    /// Inverse: re-insert whichever state the card was in (live,
+    /// archived, or — defensively — both) via `ImportEntities`, then
+    /// re-add every incident graph edge.
+    pub fn capture_inverse(&self, store: &dyn DataStore) -> KanbanResult<Vec<Command>> {
+        let live = store.get_card(self.card_id)?;
+        let archived = store.get_archived_card(self.card_id)?;
+        if live.is_none() && archived.is_none() {
+            return Err(KanbanError::not_found("Card", self.card_id));
+        }
+        let mut commands: Vec<Command> = vec![Command::Board(super::super::BoardCommand::Import(
+            super::super::ImportEntities {
+                cards: live.into_iter().collect(),
+                archived_cards: archived.into_iter().collect(),
+                ..Default::default()
+            },
+        ))];
+        let graph = store.get_graph()?;
+        let card_id = self.card_id;
+        commands.extend(super::super::dependency_commands::edges_to_undo_commands(
+            &graph,
+            |s, t| s == card_id || t == card_id,
+        ));
+        Ok(commands)
+    }
+}
+
+/// Archive one or more cards in a single command (single undo entry)
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct ArchiveCards {
+    pub ids: Vec<Uuid>,
+}
+
+impl ArchiveCards {
+    /// Inverse: one `RestoreCard` per archived card, restoring each to its
+    /// original column and position read from the live card BEFORE the
+    /// archive runs.
+    pub fn capture_inverse(&self, store: &dyn DataStore) -> KanbanResult<Vec<Command>> {
+        let mut commands: Vec<Command> = Vec::new();
+        for id in &self.ids {
+            let card = match store.get_card(*id)? {
+                Some(c) => c,
+                None => continue, // skipped (matches ArchiveCards::execute's filter)
+            };
+            commands.push(Command::Card(CardCommand::Restore(RestoreCard {
+                card_id: card.id,
+                column_id: card.column_id,
+                position: card.position,
+                timestamp: chrono::Utc::now(),
+            })));
+        }
+        Ok(commands)
+    }
+
+    pub fn execute(&self, context: &CommandContext) -> KanbanResult<()> {
+        let valid_ids = context.filter_valid_card_ids(&self.ids, "ArchiveCards");
+        if valid_ids.is_empty() && !self.ids.is_empty() {
+            return Err(KanbanError::validation(
+                "All card IDs in ArchiveCards batch are invalid",
+            ));
+        }
+        for id in &valid_ids {
+            let card = context
+                .store
+                .get_card(*id)?
+                .ok_or_else(|| KanbanError::not_found("Card", *id))?;
+            let original_column_id = card.column_id;
+            let original_position = card.position;
+            let archived = crate::ArchivedCard::new(card, original_column_id, original_position);
+            context.store.insert_archived_card(archived)?;
+            context.store.delete_card(*id)?;
+        }
+        context.store.modify_graph(Box::new(move |graph| {
+            for id in &valid_ids {
+                graph.archive_node(*id);
+            }
+            Ok(())
+        }))?;
+        Ok(())
+    }
+
+    pub fn description(&self) -> String {
+        format!("Archive {} card(s)", self.ids.len())
+    }
+}
